requirements.txt
ccxt==4.1.45
python-telegram-bot==20.6
pandas==2.1.3
numpy==1.25.2
ta==0.10.2
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
matplotlib==3.8.2
plotly==5.17.0
websocket-client==1.6.4
requests==2.31.0
pytz==2023.3
